platform/msm/ipa/ipa_v2: Mitigate usleep_range calls
Calling usleep_range with identical min and max values, results in
a delta of 0us.

This causes the issue that the scheduler gets no chance to re-use
an already scheduled interrupt for this wait which results in a
negative performance impact by adding uninterruptible sleep.

Fix the issue by allowing the delay to take 20% longer than
requested. This should be good enough for these cases.

Signed-off-by: Alex Naidis <alex.naidis@paranoidandroid.co>