clover_defconfig: Disable IP_PNP
CONFIG_IP_PNP is not set

The kernel IP autoconfiguration can only be used in practice when driver
of
the network device to be configured is also built into vmlinux. As we
build
all network device drivers as modules, there is no point keeping IP_PNP
enabled. It's completely useless on this platform.

Signed-off-by: Olaf Hering <olaf@aepfle.de>