clover_defconfig: Disable User Access Override Support
CONFIG_ARM64_UAO is not set

This config ensures user-space having access to the supplied memory.
When addr_limit is set to kernel memory, the UAO bit will be set,
allowing privileged access to kernel memory which takes a performance
hit and brings in small overhead. Therefore, disable this security
config.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>