clover_defconfig: Disable sysrq keys over serial
CONFIG_MAGIC_SYSRQ_DEFAULT_ENABLE=0x0
CONFIG_MAGIC_SYSRQ_SERIAL is not set

If the UART console is enabled in the bootloader, this could cause false
triggers, especially if the other side's UART interface has a loose TX
pin (which I've encountered).

Note that sysrq is still enabled because Android init uses it when
shutting down, and it doesn't hurt to keep it enabled for debugging.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>