audio/asoc/codecs/wcd-mbhc-adc: Report headset when it's not reported
During bootup and headset inserted, headset may first
detect as GND_MIC swap type and not report. During three
seconds polling, it's corrected as headset or headphone.
When wrk_complete is false, it will not report to userspace.
When current_plug is MBHC_PLUG_TYPE_NONE, report headset or
headphone to userspace.

Change-Id: I63394f1704c34cf355f0f4b0c33959332a69a639
Signed-off-by: Meng Wang <mengw@codeaurora.org>